add json test
